+++ app/controllers/admin/products_controller.rb
@@ -22,8 +22,6 @@ class Admin::ProductsController < ApplicationController
     end
   end
 
-
-
   def edit
     @product = Product.find(params[:id])
   end
@@ -31,7 +29,7 @@ class Admin::ProductsController < ApplicationController
   def update
     @product = Product.find(params[:id])
     if @product.update(product_params)
-      redirect_to admin_product_path
+      redirect_to admin_products_path
     else
       render :new
     end
