+++ app/controllers/carts_controller.rb
@@ -1,0 +1,2 @@
+class CartsController < ApplicationController
+end
