+++ app/controllers/products_controller.rb
@@ -13,21 +13,5 @@ class ProductsController < ApplicationController
     redirect_to :back
   end
 
-  helper_method :current_cart
-
-  def current_cart
-    @current_cart ||= find_cart
-  end
-
-  private
-
-  def find_cart
-    cart = Cart.find_by(id: session[:cart_id])
-    if cart.blank?
-      cart = Cart.create
-    end
-    session[:cart_id] = cart.id
-    return cart
-  end
 
 end
