+++ app/views/common/_navbar.html.erb
@@ -12,10 +12,10 @@
           </ul>
             <ul class="nav navbar-nav navbar-right">
                <li>
-                  <%= link_to "#" do  %>
+                  <%= link_to carts_path do  %>
                      购物车 <i class="fa fa-shopping-cart"> </i> ( <%= current_cart.products.count %> )
                   <% end %>
-               </li>              
+               </li>
                  <% if !current_user %>
                    <li><%= link_to("注册", new_user_registration_path) %> </li>
                    <li><%= link_to(content_tag(:i, '登入', class: 'fa fa-sign-in'), new_user_session_path) %></li>
